requirements.txt
pandas
numpy
yfinance
scikit-learn
xgboost==1.7.6
tensorflow-macos
tensorflow-metal
streamlit
plotly
pyyaml
requests
